# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ReelSpeed Backend API
  description: |
    Comprehensive API for ReelSpeed - AI-powered viral video generation platform.
    
    ## Features
    - 🎙️ ElevenLabs voice generation
    - 📹 Video processing and generation
    - 🔒 JWT authentication
    - 📱 Multi-platform content optimization
    - 🚀 S3/R2 asset management
    - 📊 Analytics and performance tracking
    
    ## Authentication
    Most endpoints require JWT authentication via `Authorization: Bearer <token>` header.
    
    ## Base URL
    - Development: `http://localhost:3000`
    - Production: `https://api.reelspeed.ai`
  version: 1.0.0
  contact:
    name: ReelSpeed API Support
    url: https://reelspeed.ai/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.reelspeed.ai
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/video/generate:
    post:
      tags: [Video Generation]
      summary: Generate video
      description: Start video generation with specified configuration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, input, settings]
              properties:
                type:
                  type: string
                  enum: [story, reddit, quiz, educational, viral-text]
                  example: story
                input:
                  type: object
                  properties:
                    text:
                      type: string
                    script:
                      type: string
                    title:
                      type: string
                    config:
                      type: object
                      description: Enhanced config from frontend
                settings:
                  type: object
                  properties:
                    duration:
                      type: number
                      default: 30
                    width:
                      type: number
                      default: 1080
                    height:
                      type: number
                      default: 1920
                    fps:
                      type: number
                      default: 30
                    voice:
                      type: string
                      default: default
                    background:
                      type: string
                      default: "#000000"
                    language:
                      type: string
                      default: en
                userId:
                  type: string
      responses:
        '200':
          description: Video generation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  videoId:
                    type: string
                  status:
                    type: string
                    example: processing
                  message:
                    type: string
  /api/video/status/{videoId}:
    get:
      tags: [Video Generation]
      summary: Check video generation status
      description: Get the current status of video generation
      security: []
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
          example: video_1234567890_abc123
      responses:
        '200':
          description: Video status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  videoId:
                    type: string
                  status:
                    type: string
                    enum: [processing, completed, failed]
                  progress:
                    type: number
                    minimum: 0
                    maximum: 100
                  message:
                    type: string
                  outputPath:
                    type: string
                  sizeInBytes:
                    type: number
                  duration:
                    type: number
                  error:
                    type: string
...
  /api/video/download/{filename}:
    get:
      tags: [Video Generation]
      summary: Download video file
      description: Stream/download a rendered video file with range support
      security: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          example: video_output.mp4
        - name: Range
          in: header
          schema:
            type: string
          example: bytes=0-1023
      responses:
        '200':
          description: Full video file
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '206':
          description: Partial video content (range request)
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '404':
          description: Video file not found

  # ========================
  # Asset Management Routes
  # ========================
...
  /api/assets/videos/user/{userId}:
    get:
      tags: [Asset Management]
      summary: List user videos
      description: Get user's uploaded videos from R2 storage
      security: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User videos retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  videos:
                    type: array
                    items:
                      $ref: '#/components/schemas/VideoAsset'
...
  /api/assets/list/{userId}:
    get:
      tags: [Asset Management]
      summary: List all user assets
      description: Get all assets (video, audio, images) for a user
      security: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [video, audio, image, all]
            default: all
      responses:
        '200':
          description: User assets retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  assets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
                  count:
                    type: integer

  # ========================
  # Text Video Routes
  # ========================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VideoAsset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [background, custom]
        videoUrl:
          type: string
        size:
          type: number
        duration:
          type: number
        uploadedAt:
          type: string
          format: date-time
        isPublic:
          type: boolean
...
    Asset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [video, audio, image]
        url:
          type: string
        size:
          type: number
        uploadedAt:
          type: string
          format: date-time
        isPublic:
          type: boolean
...